requirements.txt
tensorflow==2.5.0
streamlit==0.87.0
numpy==1.19.5
pillow==8.2.0
opencv-python-headless==4.5.2
